requirements.txt
langchain
langchain[openai]
langchain_core
langchain_community
langgraph
tavily-python
langchain-tavily
langgraph-checkpoint-sqlite
python-dotenv
email-validator
langchain-community
langchain-openai
streamlit
openai
psycopg2-binary
redis
fastapi
selenium
webdriver_manager